feat(state): supports `State` clone (#254)
* Set configure method

* Update interface

* Implement hasAttribute

* Implement getAttribute

* Rename file

* Configure returns instance of state store

* Implements configure

* Clone base properties

* Add tests

* Implement commit changes

* Remove configure from interface

* Fix state tests

* Fix validator tests

* Fix crypto messages tests
