+++ packages/state/source/index.ts
@@ -1,4 +1,4 @@
 export * as Attributes from "./attributes";
 export * from "./service-provider";
-export * from "./state";
+export * from "./state-store";
 export * as Wallets from "./wallets";
